requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
azure-core==1.37.0
azure-identity==1.25.1
billiard==4.2.4
bleach==6.1.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cron_descriptor==2.0.6
cryptography==46.0.3
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.58.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-otp==1.3.0
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-timezone-field==7.2.1
django_csp==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==22.0.0
frozenlist==1.8.0
geoip2==4.7.0
google-api-core==2.28.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.72.0
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.1
maxminddb==2.8.2
msal==1.34.0
msal-extensions==1.3.1
mssql-django==1.4
multidict==6.7.0
oauthlib==3.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opencensus-ext-django==0.8.0
opencensus-ext-logging==0.1.1
packaging==25.0
Pillow>=10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
#pyodbc==5.0.1
pyotp==2.9.0
pypng==0.20220715.0
pytest==9.0.2
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sentry-sdk==1.39.1
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.6.2
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.6.0
yarl==1.22.0
